Improve Kill-Server mode

+++ easytls-client-disconnect.sh
@@ -95,17 +95,17 @@ die ()
 	#exit "${2:-255}"
 	echo 'XXXXX CD - Kill Server XXXXX'
 	echo 1 > "${temp_stub}-die"
-	[ $DISABLE_KILL_PPID ] || {
-		# This always runs, except for CI
+	if [ $ENABLE_KILL_PPID ]
+	then
 		if [ $EASYTLS_FOR_WINDOWS ]
 		then
 			"${EASYTLS_PRINTF}" "%s\n%s\n" \
 				"<ERROR> ${status_msg}" "ERROR: ${1}" > "${EASYTLS_WLOG}"
-			taskkill /PID ${EASYTLS_srv_pid}
+			[ $DISABLE_KILL_PPID ] || taskkill /F /PID ${EASYTLS_srv_pid}
 		else
-			kill -15 ${EASYTLS_srv_pid}
+			[ $DISABLE_KILL_PPID ] || kill -15 ${EASYTLS_srv_pid}
 		fi
-		}
+	fi
 	exit "${2:-255}"
 }
